refactor: move func to utils
